♻️ refactor / 팀소개페이지 정적으로 변경
